pyproject.toml
[tool.poetry]
name = "django-base"
version = "0.1.0"
description = ""
authors = ["Shawn Tay <shawn.tay@genejunction.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
django-split-settings = "^1.3.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
colorlog = "^6.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
